halved memory usage for lossless compression

+++ main.cpp
@@ -28,7 +28,7 @@ extern bool quiet;
 
 int main(int argc, char* argv[])
 {
-	OCLTest<float> oclTester;
+	OCLTest<int16_t> oclTester;
 	oclTester.test();
     LogInfo("Done.\n");
